+++ tests/pkgtttest/mkprojs.py
@@ -2,8 +2,10 @@
 
 from os import makedirs
 from os.path import join
+from os.path import exists
 from subprocess import run
 from logging import debug
+from collections import namedtuple
 
 from timetracker.cmd.init import run_init_test
 
@@ -26,19 +28,36 @@ def mkdirs(tmp_home):
         makedirs(dirdoc)
     return projs
 
-def mk_projdirs(tmp_home, project='apples'):
+def mk_projdirs(tmphome, project='apples', dirgit=False):
     """Make sub-directories in a temporary directory for use in tests"""
-    pdir = join(tmp_home, f'proj/{project}')
-    dirdoc = join(pdir, 'doc')
-    makedirs(dirdoc)
-    return pdir
+    pdir = join(tmphome, f'proj/{project}')
+    makedirs(join(pdir, 'doc'))
+    if dirgit:
+        makedirs(join(pdir, '.git'))
+    return _get_expdirs(tmphome, project, dirgit)
+
+def _get_expdirs(tmphome, project='apples', dirgit=False):
+    """Make a list of expected home, project, and git directories"""
+    nto = namedtuple("ExpDirs", "dirhome dirproj dirgit dirdoc")
+    dirproj = join(tmphome, 'proj', project)
+    ntexpdirs = nto(
+        dirhome=tmphome,
+        dirproj=dirproj,
+        dirgit=join(dirproj, '.git') if dirgit else None,
+        dirdoc=join(dirproj, 'doc'))
+    for key, expdir in ntexpdirs._asdict().items():
+        print(f'{key:7} {expdir}')
+        assert exists(expdir)
+    return ntexpdirs
+
 
 def mk_projdirs_wcfgs(tmp_home, project, trksubdir='.timetracker'):
     """Make sub-directories & cfgs in a temporary directory for use in tests"""
-    dirproj = mk_projdirs(tmp_home, project)
-    fname_cfgproj = join(dirproj, trksubdir, 'config')
+    ##dirproj = mk_projdirs(tmp_home, project)
+    ntexpdirs = mk_projdirs(tmp_home, project)
+    fname_cfgproj = join(ntexpdirs.dirproj, trksubdir, 'config')
     cfgproj, cfg_global = run_init_test(fname_cfgproj, '.', project, tmp_home)
-    return dirproj, cfgproj, cfg_global
+    return ntexpdirs.dirproj, cfgproj, cfg_global
 
 
 def findhome(home):