+++ tests/test_stop.py
@@ -11,6 +11,7 @@
 #from tests.pkgtttest.mkprojs import findhome
 #from subprocess import run
 from collections import namedtuple
+from pytest import raises
 from timetracker.cmd.stop import run_stop
 
 
@@ -31,7 +32,16 @@ def test_stop():
     # 0 1
     # 1 0
     # 1 1
-    run_stop(filename_config, csvfields, quiet=quiet, keepstart=keepstart)
+    #try:
+    #    run_stop(filename_config, csvfields, quiet=quiet, keepstart=keepstart)
+    #except SystemExit as err:
+    #    print(err)
+    #print('TEST PASSED')
+    with raises(SystemExit) as excinfo:
+        run_stop(filename_config, csvfields, quiet=quiet, keepstart=keepstart)
+    assert excinfo.value.code == 0
+    print('TEST PASSED')
+
 
 if __name__ == '__main__':
     test_stop()