+++ timetracker/recorder.py
@@ -54,18 +54,15 @@ class Recorder:
     def _wr_csvhdrs(self):
         # aTimeLogger columns: Activity From To Notes
         with open(self.csv, 'w', encoding='utf8') as prt:
-            prt.write('start_day',
+            prt.write('start_day,'
                       'start_datetime,'
                       'start_tic,'
-                      # Stop
-                      'stop_day',
+                      'stop_day,'
                       'stop_datetime,'
                       'stop_toc,'
-                      # Duration
                       'duration,'
-                      # Info
-                      'tags',
-                      'type',
+                      'tags,'
+                      'type,'
                       'comment\n')
 
     def _rd_datetime(self):