+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Summary
 * [**Unreleased**](#unreleased)
+* **Release 2025-02-08 v0.2a1** Cleaned up tests; Added pytest to GitHub workflows upon push
 * **Release 2025-02-07 v0.2a0** Added functionality for running in project subdirs
 * [**Release 2025-02-05 v0.1a8**](#release-2025-02-05-v01a8) Added DOI for citing
 * [**Release 2025-02-02 v0.1a4**](#release-2025-02-02-v01a4) Added global config file